Export PublicElb URL To Allow CrossStack Reference
When creating the stack, and you define a domain for use with SSL, the current Output value of `TeamCityServerUrl` references the public ELB DNS address. This is fine for access but breaks most browsers since when accessing `teamc-publi-xxxxx-xxxxx.us-west-x.elb.amazonaws.com` SSL breaks since the domain you are requesting is not what the SSL cert is using.

In our case, all we need to do is add a Route 53 Alias in the same HostedZone we used for the domain of the SSL cert and the SSL "issues" go away.

To do this dynamically we would need the ELB DNS name to Exported such that we can import into our existing stacks.

Does this make sense?

Thanks in advance.